// dashboard/ui/src/api/generated/models/engine.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MDDash Dashboard API
 * API for experiment, simulation, compute, analysis, notebook, file, and publication workflows.
 * OpenAPI spec version: 1.0.0
 */

export type Engine = typeof Engine[keyof typeof Engine];


export const Engine = {
  GMX: 'GMX',
  AMBER: 'AMBER',
} as const;