// dashboard/ui/src/api/generated/models/amberJobRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MDDash Dashboard API
 * API for experiment, simulation, compute, analysis, notebook, file, and publication workflows.
 * OpenAPI spec version: 1.0.0
 */

export interface AmberJobRequest {
  binary: string;
  ewald: string;
  /** @minimum 1 */
  np: number;
  /** @minimum 1 */
  ntomp: number;
}