// dashboard/ui/src/api/generated/models/gromacsJobRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MDDash Dashboard API
 * API for experiment, simulation, compute, analysis, notebook, file, and publication workflows.
 * OpenAPI spec version: 1.0.0
 */
import type { GromacsJobRequestNb } from './gromacsJobRequestNb.ts';
import type { GromacsJobRequestPme } from './gromacsJobRequestPme.ts';

export interface GromacsJobRequest {
  /** @minimum 1 */
  np: number;
  /** @minimum 1 */
  ntomp: number;
  pme: GromacsJobRequestPme;
  nb: GromacsJobRequestNb;
}